% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf_plots.R
\name{acf_plots}
\alias{acf_plots}
\title{Create Auto Correlation Plots for Models}
\usage{
acf_plots(input, columns, pre_model = FALSE, stan = FALSE)
}
\arguments{
\item{input}{list List of the model output}

\item{columns}{vector The target columns to assess}

\item{pre_model}{bool If you are passing the pre-model}

\item{stan}{bool If you are using the stan model}
}
\value{
plotted objects
}
\description{
Create auto correlation plots for the models to test if there
is any high autocorrelation in the sampling space, which would
imply that sampler should be adjusted or the data needs to be
adjusted.
}
