% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_model.R
\name{create_model_string}
\alias{create_model_string}
\title{Define a Model that is JAGS Usable}
\usage{
create_model_string()
}
\value{
Full path to the model specification
}
\description{
Create the model string, save it to a temporary folder,
and return back the location of the temporary model
file for usage by JAGS later.
}
\examples{
create_model_string()
}
