% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_stan_model.R
\name{create_stan_model_string}
\alias{create_stan_model_string}
\title{Define the Pre-Model Using Stan}
\usage{
create_stan_model_string()
}
\value{
Full path to the model specification
}
\description{
Create the model string, save it to a temporary folder,
and return back the location of the temporary model
file for usage by Stan later. This is experimental and has
been validated, but caution should be used when utilizing
the Stan backend as the pre-model.
}
\examples{
create_stan_model_string()
}
