% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_model.R
\name{multivariate_model}
\alias{multivariate_model}
\title{Model the Data with Multivariate Adjustment}
\usage{
multivariate_model(
  formula,
  data,
  columns,
  jags_model,
  a_c_matrix,
  n_burn = 1000,
  n_thin = 1,
  n_steps = 10000,
  seed = 42,
  univariate = FALSE,
  sds = NULL,
  variances = NULL
)
}
\arguments{
\item{formula}{character The model formula}

\item{data}{data.frame The data to model}

\item{columns}{vector The columns that are relevant}

\item{jags_model}{jgas.model A fitted pre-model}

\item{a_c_matrix}{matrix The attenuation-contamination matrix}

\item{n_burn}{numeric Number of draws to burn at the start, default = 1000}

\item{n_thin}{numeric Thinning factor for the draws, default = 1}

\item{n_steps}{numeric The total number of draws to run, default = 10,000}

\item{seed}{numeric The random seed to set}

\item{univariate}{bool Whether or not to run the univariate model,
default = TRUE}

\item{sds}{list If you are running the univariate model, the listing of
standard deviations needs to be applied}

\item{variances}{list If you are running the univariate model, the listing of
variances need to be applied}
}
\value{
List with fitted models ready for further analysis
}
\description{
Using the methodology generated by Muoko et. al (see README for
full citation), run the modelling with the JAGS sampler. This model
uses the calculated attenuation-contamination matrix to adjust
the various covariates needed in the model. The function accepts
the computed pre-model, the attenuation-contamination coefficient,,
a model string, and standard deviations. There is an optional
argument for also fitting the univariate model to run further
comparisons between the naive, univariate, and multivariate models
in later steps.
}
