% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_coefficients.R
\name{normalize_coefficients}
\alias{normalize_coefficients}
\title{Normalize the upper and lower bounds of a validity coefficient}
\usage{
normalize_coefficients(n, lower_bound, upper_bound, interval = 0.95)
}
\arguments{
\item{n}{the samples for the normal distribution}

\item{lower_bound}{numeric the lower boundary of the validity coefficient}

\item{upper_bound}{numeric the upper boudnary of the validity coefficient}

\item{interval}{numeric the confidence interval to use (default 0.95)}
}
\value{
Normal distribution
}
\description{
Using the Fisher Z Transformation, normalize the upper and lower
boundary of a validity coefficient to generate a Normal distribution
based on the mean and standard deviation. The default behavior is
to assume a 95\% CI, but this can also be specified via the \code{interval}
parameter if 99\% or 90\% CI is preferred.
}
\examples{
n <- 1000
coef_upper <- 0.9
coef_lower <- 0.3
ci <- 0.95
normalize_coefficients(n, coef_lower, coef_upper, ci)
}
