% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_diagnostics.R
\name{plot_a_list}
\alias{plot_a_list}
\title{Custom Function to Use Patchwork}
\usage{
plot_a_list(plot_list, rows, columns)
}
\arguments{
\item{plot_list}{list List of plots}

\item{rows}{numeric Number of rows}

\item{columns}{numeric Number of columns}
}
\description{
Uses patchwork to make gridding easier for plotting the
traceplots and acf plots for the Stan pre-model.
}
