% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_plot.R
\name{plot_simulation}
\alias{plot_simulation}
\title{Plotting code for the simulation results}
\usage{
plot_simulation(data)
}
\arguments{
\item{data}{list(data.frame) The output from the simulation}
}
\value{
list The plot objects
}
\description{
Creates density plots of the means and SD of the results
from the simulation study. These will be facet grid
ggplot2 objects.
}
