% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation_study}
\alias{simulation_study}
\title{Run a simulations of the measurement error model}
\usage{
simulation_study(
  parameters,
  data,
  formula,
  columns,
  stan = FALSE,
  seed = 42,
  draws = 20
)
}
\arguments{
\item{parameters}{vector The validity parameters to use}

\item{data}{data.frame The data to use for the simulation analysis}

\item{formula}{character The formula for the model}

\item{columns}{vector The columns within the data for the covariates}

\item{stan}{boolean Whether or not to run with the Stan backend}

\item{seed}{numeric The random seed to set}

\item{draws}{numeric The number of simulations to run}
}
\value{
list with the means and the SDs of the parameters
}
\description{
Create a simulation analysis based on a set of parameters
with varied data. The input to this function will be a dataframe
and a set of parameters that will then be used to generate
synthetic forms of the data. You can also specify how many
different "draws" of the data you want to pull from. Parallel
computing will be used, so again be judicious with how many
draws you want to pull.
}
\details{
As with the sensitivity analysis, be very careful when using
this function as parallel computing is expensive and difficult
to debug. So, ultimately, buyer beware.
}
