% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize_with_return}
\alias{standardize_with_return}
\title{Standardize the data of 1-D vector}
\usage{
standardize_with_return(data)
}
\arguments{
\item{data}{vector The vector to standardize}
}
\value{
List with the original standard deviation, mean,
and the standardized data
}
\description{
Perform standardization of data on a 2-D dataframe
type object. Standardization in this case refers
to (x - mean(x)) / sd(x) where X is a 1-dimensional
vector.
}
\examples{
data <- c(1, 2, 3, 4)
standardize_with_return(data)
}
