% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attenuation_contamination_matrix.R
\name{attenuation_matrix}
\alias{attenuation_matrix}
\title{Create Attenuation Contamination Matrix}
\usage{
attenuation_matrix(model_output, columns, validity_coefficients, stan = FALSE)
}
\arguments{
\item{model_output}{list List of the mcmc summary and covariance matrix}

\item{columns}{vector Vector of the column names that are being assessed}

\item{validity_coefficients}{vector Vector of the validity coefficients}

\item{stan}{boolean If you are passing in a Stan backend pre-model}
}
\value{
List with the attenuation-contamination matrix and the standard
deviations
}
\description{
From the output JAGS model, create the needed attenuation
contamination matrix for further analysis.
}
