% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipeline}
\alias{pipeline}
\title{Pipeline used for running a model start to finish.}
\usage{
pipeline(data, formula, parameters, columns, stan = FALSE, seed = 42)
}
\arguments{
\item{data}{data.frame The input data}

\item{formula}{character The formula for the multivariate model}

\item{parameters}{vector The validity coefficients}

\item{columns}{vector The columns of the covariates}

\item{stan}{bool If you would like to run with the Stan backend}

\item{seed}{numeric The random seed to use}
}
\description{
This is an unexported function that can run the entire
pre-model, attenuation-contamination matrix, and model
fitting from start to finish.
}
