% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_covariates}
\alias{plot_covariates}
\title{Diagnostics for Models}
\usage{
plot_covariates(model_output, columns)
}
\arguments{
\item{model_output}{list List of the model output}

\item{columns}{vector The target columns to assess}
}
\value{
plotted objects
}
\description{
Using the previously fit models, add the diagnostics for
assessing the adjusted covariates based on the naive,
univariate, and multivariate models.
}
