% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceplots.R
\name{traceplots}
\alias{traceplots}
\title{Create Traceplots for the Parameters}
\usage{
traceplots(input, columns, pre_model = FALSE, stan = FALSE)
}
\arguments{
\item{input}{list List of the model output}

\item{columns}{vector The target columns to assess}

\item{pre_model}{bool If you are passing the pre-model}

\item{stan}{bool If the model is stan based}
}
\value{
plotted objects
}
\description{
Create sampler traceplots to run diagnostics on whether or
not the sampler converged when creating the posterior. Columns
should always be in the order that they appear in the model.
}
